:building_construction: Restructuring of domain and infrastructure layers

+++ ManageFood.Domain.Entities/FoodShop/InventoryEntity.cs
@@ -7,7 +7,7 @@ namespace ManageFood.Domain.Entities.FoodShop
     public required Guid ProductId { get; set; }
     public required int Quantity { get; set; }
     public required int QuantityAvailable { get; set; }
-    public required float Unit {  get; set; }
+    public required decimal Unit {  get; set; }
     public required UnitType UnitType { get; set; }
     public required decimal Price { get; set; }
     public DateTimeOffset Created { get; set; }
